#263 Image cdo path with spaces are not properly transformed
The project name stored in description and workspace path are decoded,
but the cdo path are encoded so it is needed to encode the project name
to transform the cdo path.

In addition,
org.eclipse.emf.diffmerge.sirius.SiriusScope.siriusImageHelper is
initialized to better test SiriusScope independently of the editor that
uses IComparison.

Bug: https://github.com/eclipse/org.eclipse.emf.diffmerge.core/issues/263
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>